+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to Rust's notion of
     `bellman::multicore::Waiter<Result<G, SynthesisError>>` instead of
     `Box<dyn Future<Item = G, Error = SynthesisError>>`.
   - `bellman::multicore::log_num_cpus` is renamed to `log_num_threads`.
+- `bellman::multiexp::SourceBuilder::new` is renamed to `SourceBuilder::build`.
 
 ### Removed
 - `bellman::multicore::WorkerFuture` (replaced by `Waiter`).